requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-nose==1.4.6
django-rest-swagger==2.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jsonnet==0.18.0
MarkupSafe==2.1.3
nose==1.3.7
openapi-codec==1.3.2
outcome==1.3.0.post0
psycopg2==2.9.4
pycryptodome==3.15.0
pynose==1.4.8
PySocks==1.7.1
pytz==2023.3.post1
requests==2.28.1
selenium==4.7.2
simplejson==3.19.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
wsproto==1.2.0
django-allauth
django-sslserver
